requirements/requirements.txt
apiritif==0.3
astunparse==1.5.0
bzt==1.10.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
colorlog==3.1.0
cssselect==1.0.1
decorator==4.1.2
EasyProcess==0.2.3
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jsonpath-rw==1.4.0
locustio==0.8.1
lxml==4.1.1
MarkupSafe==1.0
msgpack-python==0.4.8
nose==1.3.7
ply==3.10
progressbar33==2.4
psutil==5.4.1
py==1.4.34
pytest==3.2.3
PyVirtualDisplay==0.2.1
PyYAML==3.12
pyzmq==16.0.3
requests==2.18.4
selenium==3.7.0
six==1.11.0
urllib3==1.22
urwid==1.3.1
Werkzeug==0.12.2
